requirements.txt
# Core dependencies
faiss-cpu>=1.7.0
numpy>=1.21.0
pandas>=1.3.0
sqlalchemy>=1.4.0
openai>=0.27.0

# Database drivers
pymysql>=1.0.0
psycopg2-binary>=2.9.0

# Utility libraries
python-dotenv>=0.19.0
requests>=2.28.0
pydantic>=1.10.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
